# trace: replace newEvictedQueue sync.OnceFunc with sync.Once